requirements.txt
streamlit
pandas
scikit-learn
lightgbm
google-generativeai
python-dotenv
requests
elevenlabs